pyproject.toml
[tool.poetry]
name = "agi"
version = "0.1.0"
description = ""
authors = ["gregology", "ryoung2512", "xiangshiyin"]
readme = "README.md"
packages = [{include = "agi", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.35.4"
typer = "^0.12.3"
jinja2 = "^3.1.4"
neo4j = "^5.21.0"
python-dotenv = "^1.0.1"
podman-compose = "^1.1.0"
flake8 = "^7.1.0"
black = "^24.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
vcrpy = "^6.0.1"
deepdiff = "^7.0.1"

[tool.poetry.scripts]
agi = "agi.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
